refactor: move all static web content into marketing folder
