+++ src/app/(marketing)/about/components/about-carousel.tsx
