+++ src/app/(marketing)/about/components/additional-section.tsx
