+++ src/app/(marketing)/about/components/hero-section.tsx
