+++ src/app/(marketing)/about/components/values-section.tsx
