+++ src/app/(marketing)/about/page.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { ValuesSection } from '@/app/about/components/values-section';
-import { AdditionalSection } from '@/app/about/components/additional-section';
+import { ValuesSection } from '@/app/(marketing)/about/components/values-section';
+import { AdditionalSection } from '@/app/(marketing)/about/components/additional-section';
 import { aboutConfig } from '@/config/about';
-import { HeroSection } from '@/app/about/components/hero-section';
+import { HeroSection } from '@/app/(marketing)/about/components/hero-section';
 import parseTextWithMarkup from '@/config/parser/parseTextWithMarkup';
 
 export default function AboutPage() {
