+++ src/app/(marketing)/enterprise-solutions/components/get-in-touch.tsx
@@ -13,8 +13,8 @@ import {
 } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
 import { toast } from '@/components/ui/use-toast';
-import { Textarea } from '../../../components/ui/textarea';
-import { Checkbox } from '../../../components/ui/checkbox';
+import { Textarea } from '@/components/ui/textarea';
+import { Checkbox } from '@/components/ui/checkbox';
 import { enterpriseSolutionsConfig } from '@/config/enterprise-solutions';
 
 const FormSchema = z.object({
