+++ src/app/(marketing)/knowledge-hub/components/blog-section.tsx
