+++ src/app/(marketing)/knowledge-hub/components/hero-section.tsx
