+++ src/app/(marketing)/knowledge-hub/components/video-section.tsx
