+++ src/app/(marketing)/legal/components/legal-content-section.tsx
