+++ src/app/(marketing)/legal/components/legal-hero-section.tsx
