+++ src/app/(marketing)/legal/components/legal-section-card.tsx
