+++ src/app/(marketing)/legal/page.tsx
